Properly handle frozen strings
This rewrites the ACL parsing in the type.

First of all, it renames variables from a single letter to multiple so
it properly describes what they are.

It also fixes the octal parsing, which always resulted in rwx.

More importantly, it fixes the frozen string behavior. It does this by
using += instead of << and avoiding sub!.

Fixes: a252009959434a6eb29400368d112fdfcdad927d